refactor: carRepository, carRacingGameEngine 제거 후 Cars 일급 컬렉션으로 대체
