+++ src/main/java/racingcar/domain/car/GoForwardOrStop.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 public class GoForwardOrStop {
     private static final int MIN_GO_FORWARD_VALUE = 3;
