+++ src/main/java/racingcar/domain/car/Position.java
@@ -38,4 +38,5 @@ public class Position {
     public int hashCode() {
         return Objects.hash(position);
     }
+
 }
